ipc: linux: Support incremental allowed ips updates
[1] adds support for the WGALLOWEDIP_F_REMOVE_ME flag to WireGuard's
Linux driver which is a direct way of removing a single allowed ip from
a peer. Extend the interface of `wg set` to leverage this feature,
allowing for incremental updates to a peer's configuration. By default,
allowed-ips fully replaces a peer's allowed ips using
WGPEER_REPLACE_ALLOWEDIPS under the hood. When '+' or '-' is prepended
to any ip in the list, wg clears WGPEER_F_REPLACE_ALLOWEDIPS and sets
the WGALLOWEDIP_F_REMOVE_ME flag on any ip prefixed with '-'.

$ wg set wg0 peer <PUBKEY> allowed-ips +192.168.88.0/24,-192.168.0.1/32

This command means "add 192.168.88.0/24 to this peer's allowed ips if
not present, and remove 192.168.0.1/32 if present".

Use -isystem so that headers in uapi/ take precedence over system
headers; otherwise, the build will fail on Linux systems without updated
headers.

[1]: https://lore.kernel.org/netdev/20240905200551.4099064-1-jrife@google.com/

Signed-off-by: Jordan Rife <jordan@jrife.io>